Problem 238 - Product of Array Except Self
I was able to tackle this problem with my mentor Andy. He was able to let me think about the problem on my own while steering me in the right direction. He was able to show me how to solve the problem without coding the solution for me. The apporach we used is the Prefxi Sum Approach, using a for loop to populate the prefix array going forward in the nums array and then using a backwards for loop to populate the suffix array. Using these two arrays, multiply them to the correspodning indexes to get the desired output.