121 - Best Time to Buy and Sell Stock
This was after my first actual attempt of the problem, had the right idea but did not execute and implement the solution optimally until watching an exlpanation to the two pointer solution. Is two pointer and sliding window the same? Neetcode's solution is advertised as sliding window but his explanation describes it as two pointer, interesting