347 - Top K Frequent Elements
This was my first time looking at the problem. I was able to get an idea of how to approach the problem using a Hashmap, and was able to pass 9 of 21 testcases. But then after looking at the solution on Neetcode, I was able to decipher what I was missing and find a more optimal solution to the problem. Will come back to this problem