238 Product of Array Except Self (second look)
I was able to try and solve this problem again but was not able to solve on my own. I looked at the solution and learned about utilizing prefix and postfix to solve this problem. I also reviewed the optimal solution and it makes sense on paper but is hard to implement. I also attempted to pass the problem using a division operator just to try and was able to pass some test cases