242 - Valid Anagram
Finished up this problem, after looking at the solution and learning the solution. I learned that using an object and using a Map() is the same time complexity and can be used the same with different syntax. I ultimately learned the object solution as it has less busy words in the solution when compared to using the built-in Map(). Onto the next one.